pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ragforge"
version = "0.1.0"
description = "A zero-configuration, opinionated RAG and GraphRAG pipeline with Neo4j integration."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "groq",
    "qdrant-client",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv",
    "fastembed>=0.7.4",
    "neo4j>=5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ragforge*"]

[project.urls]
"Homepage" = "https://github.com/example/ragforge"
